+++ IndiaEvents.Models/Models/RequestSheets/EventRequestExpenseSheet.cs
@@ -18,8 +18,8 @@
         public string? EventId { get; set; }
         public string? ExpenseType { get; set; }
         //public string? Amount { get; set; }
-        public int? AmountIncludingTax { get; set; }
-        public int? ExcludingTaxAmount { get; set; }
+        public double? AmountIncludingTax { get; set; }
+        public double? ExcludingTaxAmount { get; set; }
         public string? BtcorBte { get; set; }
         public string? EventTopic { get; set; }
         public string? EventType { get; set; }
